prefect dev & publish scripts
