+++ package.json
@@ -3,13 +3,11 @@
   "version": "1.0.0",
   "description": "Multiplayer code editor system",
   "main": "src/index.ts",
-  "type": "module",
-  "bin_": "src/server/index.js",
-  "bin_build_server": "dist_server/server/index.js",
+  "type": "module", 
   "bin": "dist_server/server/index.js",
   "files": [
-    "./dist/**/*",
-    "./space/**/*"
+    "dist/**",
+    "src/**"
   ],
   "scripts": {
     "test": "vitest run",
@@ -21,22 +19,16 @@
     "build-release": "vite build --mode release",
     "preview": "concurrently \"npm run test-interactive\" \"vite preview\"",
     "prepublishOnly": "npm run build-release",
-    "start-server": "node  src/server/index.js --space=./space --site=./dist",
-    "start-server_n": "node  src/server/index.js",
+    "start-server": "node  src/server/index.js dev --space=./space --site=./dist", 
     "build-server": "npx babel ./src --out-dir dist_server",
     "pkg": "pkg . ",
-    "pkg-pack": "copyfiles ./dist/* ./dist/**/* ./dist_bin  & copyfiles ./space/* ./dist_bin ",
+    "pkg-pack": "copyfiles ./dist/* ./dist/**/* ./dist_server_bin  & copyfiles ./space/* ./dist_server_bin ",
     "pkg-go": "concurrently \"npm run pkg\" \"npm run pkg-pack\""
   },
   "pkg": {
     "scripts": "dist_server/server/index.js",
-    "assets": [
-      "space/**/*.*",
-      "space/hello.md",
-      "dist/**/*.*",
-      "dist/hello.md"
-    ],
-    "outputPath": "dist_bin"
+    "assets": [],
+    "outputPath": "dist_server_bin"
   },
   "repository": {
     "type": "git",
@@ -168,6 +160,7 @@
     "@vitejs/plugin-react": "^4.2.1",
     "babel-plugin-transform-import-meta": "^2.2.1",
     "concurrently": "^8.2.2",
+    "copy-dir": "^1.3.0",
     "copyfiles": "^2.4.1",
     "prettier": "^3.2.5",
     "sass": "^1.76.0",
