+++ src/server/cmdArg.js
@@ -1,12 +1,18 @@
 import path from 'path';
-import fs from 'fs';
-import { isDirectory } from './isDirectory';
-import { AppEnv } from '../appEnv';
+import fs from 'fs'; 
+ 
+
+// 
+const isDevMode = function(){ 
+  const dev = process.argv.find((arg) =>
+    arg === 'dev',
+  );
+  return dev;
+}
 
 const getCurrentExeFolder = function(){ 
   //process.cwd();////__dirname;// path.dirname(process.execPath);
-  const execBin = process.argv[0];
-  var result = path.dirname(execBin);
+  const result = isDevMode()? process.cwd(): path.dirname(process.argv[0]); 
    
   console.log(`getCurrentExeFolder:${result}`)
   return result; 
@@ -74,39 +80,32 @@ export const getDocumentSpaceFromArgs = function (defaultVal = "space") {
     return result;
 }
 
-const cp = function(from, to){
-  try {
-    console.log(`try cp: ${from} => ${to} `)
-    fs.cpSync(from, to, {
-      recursive: true,
-    });
-  } catch (error) {
-    console.error(error.message);
-  }
-}
 
- 
-const ls = async (dirName) => {
-  console.log('try ls:' + dirName);
-  const getFileList = async (dirName) => {
-    const files = await fs.readdir(dirName);
-
-    return files;
-  };
-
-  getFileList(dirName).then((files) => {
-      console.log(files);
-  });
-};
 
  
 export const prepareSpaceByAsset = function(assetName,toSpace){
-  const to = path.join(toSpace,assetName);
-  if (isDirectory(to)){
-    return ;
-  }
-  //ls(AppEnv.getProjectRootPath());
-  const from =  path.join(AppEnv.getProjectRootPath(), assetName) 
+  // const to = path.join(toSpace,assetName);
   
-  cp(from,to);
+  // //ls(AppEnv.getProjectRootPath());
+  // const from =  path.join(AppEnv.getProjectRootPath(), assetName) 
+  
+  // cp(from,to);
+}
+
+
+export const AppEnvNoSense ={
+  /*
+   /snapshot/vzcode/dist_server/server/ => /snapshot/vzcode/
+  */
+  getProjectRootPath(){
+      let currentPath = __dirname;
+      
+      const list = currentPath.split(path.sep);
+      var result = "";
+      for (let index = 0; index < list.length -2; index++) { 
+          result = path.join(result,list[index]);
+      }
+      console.log(`🔥 getProjectRootPath: ${result}`)
+      return result;
+  }
 }
