+++ src/server/isDirectory.js
@@ -1,1 +1,3 @@
-export const isDirectory = (file) => file.endsWith('/');
+import path from 'path';
+// export const isDirectory = (file) => file.endsWith('/');
+export const isDirectory = (file) => file.endsWith(path.sep);
