--- src/server/main.py
@@ -1,1 +1,0 @@
-await
